+++ packages/driver/cypress/e2e/commands/actions/check.cy.ts
@@ -221,7 +221,7 @@ describe('src/cy/commands/actions/check', () => {
       cy.get(':checkbox:first').check({ scrollBehavior: 'bottom' })
 
       cy.get(':checkbox:first').then((el) => {
-        expect(el[0].scrollIntoView).to.be.calledWith({ block: 'end', inline: 'end' })
+        expect(el[0].scrollIntoView).to.be.calledWith({ block: 'end' })
       })
     })
 
@@ -235,68 +235,6 @@ describe('src/cy/commands/actions/check', () => {
 
       cy.get(':checkbox:first').then((el) => {
         expect(el[0].scrollIntoView).not.to.be.called
-      })
-    })
-
-    it('can specify scrollBehavior bottom in config', { scrollBehavior: 'bottom' }, () => {
-      cy.get(':checkbox:first').then((el) => {
-        cy.spy(el[0], 'scrollIntoView')
-      })
-
-      cy.get(':checkbox:first').check()
-
-      cy.get(':checkbox:first').then((el) => {
-        expect(el[0].scrollIntoView).to.be.calledWith({ block: 'end', inline: 'end' })
-      })
-    })
-
-    it('can specify scrollBehavior center in config', { scrollBehavior: 'center' }, () => {
-      cy.get(':checkbox:first').then((el) => {
-        cy.spy(el[0], 'scrollIntoView')
-      })
-
-      cy.get(':checkbox:first').check()
-
-      cy.get(':checkbox:first').then((el) => {
-        expect(el[0].scrollIntoView).to.be.calledWith({ block: 'center', inline: 'center' })
-      })
-    })
-
-    it('can specify scrollBehavior nearest in config', { scrollBehavior: 'nearest' }, () => {
-      cy.get(':checkbox:first').then((el) => {
-        cy.spy(el[0], 'scrollIntoView')
-      })
-
-      cy.get(':checkbox:first').check()
-
-      cy.get(':checkbox:first').then((el) => {
-        expect(el[0].scrollIntoView).to.be.calledWith({ block: 'nearest', inline: 'nearest' })
-      })
-    })
-
-    it('does not scroll when scrollBehavior is false in config', { scrollBehavior: false }, () => {
-      cy.get(':checkbox:first').scrollIntoView()
-      cy.get(':checkbox:first').then((el) => {
-        cy.spy(el[0], 'scrollIntoView')
-      })
-
-      cy.get(':checkbox:first').check()
-
-      cy.get(':checkbox:first').then((el) => {
-        expect(el[0].scrollIntoView).not.to.be.called
-      })
-    })
-
-    it('calls scrollIntoView by default', () => {
-      cy.scrollTo('top')
-      cy.get(':checkbox:first').then((el) => {
-        cy.spy(el[0], 'scrollIntoView')
-      })
-
-      cy.get(':checkbox:first').check()
-
-      cy.get(':checkbox:first').then((el) => {
-        expect(el[0].scrollIntoView).to.be.calledWith({ block: 'start', inline: 'start' })
       })
     })