+++ packages/driver/cypress/e2e/commands/actions/trigger.cy.ts
@@ -608,7 +608,7 @@ describe('src/cy/commands/actions/trigger', () => {
         cy.get('button:first').trigger('mouseover', { scrollBehavior: 'bottom' })
 
         cy.get('button:first').then((el) => {
-          expect(el[0].scrollIntoView).to.be.calledWith({ block: 'end', inline: 'end' })
+          expect(el[0].scrollIntoView).to.be.calledWith({ block: 'end' })
         })
       })
 
@@ -622,68 +622,6 @@ describe('src/cy/commands/actions/trigger', () => {
 
         cy.get('button:first').then((el) => {
           expect(el[0].scrollIntoView).not.to.be.called
-        })
-      })
-
-      it('can specify scrollBehavior bottom in config', { scrollBehavior: 'bottom' }, () => {
-        cy.get('button:first').then((el) => {
-          cy.spy(el[0], 'scrollIntoView')
-        })
-
-        cy.get('button:first').trigger('mouseover')
-
-        cy.get('button:first').then((el) => {
-          expect(el[0].scrollIntoView).to.be.calledWith({ block: 'end', inline: 'end' })
-        })
-      })
-
-      it('can specify scrollBehavior center in config', { scrollBehavior: 'center' }, () => {
-        cy.get('button:first').then((el) => {
-          cy.spy(el[0], 'scrollIntoView')
-        })
-
-        cy.get('button:first').trigger('mouseover')
-
-        cy.get('button:first').then((el) => {
-          expect(el[0].scrollIntoView).to.be.calledWith({ block: 'center', inline: 'center' })
-        })
-      })
-
-      it('can specify scrollBehavior nearest in config', { scrollBehavior: 'nearest' }, () => {
-        cy.get('button:first').then((el) => {
-          cy.spy(el[0], 'scrollIntoView')
-        })
-
-        cy.get('button:first').trigger('mouseover')
-
-        cy.get('button:first').then((el) => {
-          expect(el[0].scrollIntoView).to.be.calledWith({ block: 'nearest', inline: 'nearest' })
-        })
-      })
-
-      it('does not scroll when scrollBehavior is false in config', { scrollBehavior: false }, () => {
-        cy.scrollTo('top')
-        cy.get('button:first').then((el) => {
-          cy.spy(el[0], 'scrollIntoView')
-        })
-
-        cy.get('button:first').trigger('mouseover')
-
-        cy.get('button:first').then((el) => {
-          expect(el[0].scrollIntoView).not.to.be.called
-        })
-      })
-
-      it('calls scrollIntoView by default', () => {
-        cy.scrollTo('top')
-        cy.get('button:first').then((el) => {
-          cy.spy(el[0], 'scrollIntoView')
-        })
-
-        cy.get('button:first').trigger('mouseover')
-
-        cy.get('button:first').then((el) => {
-          expect(el[0].scrollIntoView).to.be.calledWith({ block: 'start', inline: 'start' })
         })
       })